# api/swagger.yaml
basePath: /
definitions:
...
  model.ArticleItem:
    properties:
      id:
        type: integer
      image:
        type: string
      is_advertorial:
        type: boolean
      is_spoiler:
        type: boolean
      summary:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      username:
        type: string
      view:
        type: integer
    type: object
...
  model.Date:
    properties:
      day:
        type: integer
      month:
        type: integer
      year:
        type: integer
    type: object
...
  model.NewsItem:
    properties:
      comment:
        type: integer
      content:
        type: string
      date:
        type: string
      forum_id:
        type: integer
      id:
        type: integer
      image:
        type: string
      title:
        type: string
      username:
        type: string
    type: object
...
  model.People:
    properties:
      alternativeNames:
        items:
          type: string
        type: array
      birthday:
        $ref: '#/definitions/model.Date'
      familyName:
        type: string
      favorite:
        type: integer
      givenName:
        type: string
      id:
        type: integer
      image:
        type: string
      more:
        type: string
      name:
        type: string
      website:
        type: string
    type: object
  model.PeopleCharacter:
    properties:
      anime:
        $ref: '#/definitions/model.Role'
      character:
        $ref: '#/definitions/model.Role'
    type: object
...
  model.Recommendation:
    properties:
      recommended:
        $ref: '#/definitions/model.Source'
      source:
        $ref: '#/definitions/model.Source'
      users:
        items:
          $ref: '#/definitions/model.RecommendationUser'
        type: array
    type: object
  model.RecommendationUser:
    properties:
      content:
        type: string
      username:
        type: string
    type: object
...
  model.Role:
    properties:
      id:
        type: integer
      image:
        type: string
      name:
        type: string
      role:
        type: string
    type: object
...
  model.Source:
    properties:
      id:
        type: integer
      image:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
...
  utils.Response:
    properties:
      data:
        type: object
      message:
        type: string
      meta:
        type: object
      status:
        type: integer
    type: object
info:
  contact:
    email: axel.rl.404@gmail.com
    name: Axel
    url: https://github.com/rl404
  description: Scraping/parsing MyAnimeList website to a useful and easy-to-use data.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Malscraper API
paths:
...
  /people/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: People ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.People'
              type: object
      summary: Get people details
      tags:
      - people
  /people/{id}/article:
    get:
      consumes:
      - application/json
      parameters:
      - description: People ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.ArticleItem'
                  type: array
              type: object
      summary: Get people featured article list
      tags:
      - people
  /people/{id}/characters:
    get:
      consumes:
      - application/json
      parameters:
      - description: People ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.PeopleCharacter'
                  type: array
              type: object
      summary: Get people anime character role list
      tags:
      - people
  /people/{id}/manga:
    get:
      consumes:
      - application/json
      parameters:
      - description: People ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Role'
                  type: array
              type: object
      summary: Get people published manga list
      tags:
      - people
  /people/{id}/news:
    get:
      consumes:
      - application/json
      parameters:
      - description: People ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.NewsItem'
                  type: array
              type: object
      summary: Get people news list
      tags:
      - people
  /people/{id}/pictures:
    get:
      consumes:
      - application/json
      parameters:
      - description: People ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    type: string
                  type: array
              type: object
      summary: Get people picture list
      tags:
      - people
  /people/{id}/staff:
    get:
      consumes:
      - application/json
      parameters:
      - description: People ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Role'
                  type: array
              type: object
      summary: Get people anime staff role list
      tags:
      - people
...
  /recommendation/anime/{id1}/{id2}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Anime ID
        in: path
        name: id1
        required: true
        type: integer
      - description: Anime ID
        in: path
        name: id2
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Recommendation'
              type: object
      summary: Get anime recommendation from specific anime.
      tags:
      - recommendation
  /recommendation/manga/{id1}/{id2}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Manga ID
        in: path
        name: id1
        required: true
        type: integer
      - description: Manga ID
        in: path
        name: id2
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Recommendation'
              type: object
      summary: Get manga recommendation from specific manga.
      tags:
      - recommendation
  /recommendations/anime:
    get:
      consumes:
      - application/json
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Recommendation'
                  type: array
              type: object
      summary: Get anime recommendation list.
      tags:
      - recommendation
  /recommendations/manga:
    get:
      consumes:
      - application/json
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Recommendation'
                  type: array
              type: object
      summary: Get manga recommendation list.
      tags:
      - recommendation
...
schemes:
- http
- https
swagger: "2.0"